+++ daily_summary/main.py
@@ -1,10 +1,126 @@
 import argparse
 from datetime import datetime
-from .generate_daily_summary import generate
-from .summarize_diff import summarize_all_diffs
-from .git_data_extract import extract_git_data
+import logging
+import git
+from tqdm import tqdm
 import os
 import sys
+from openai import OpenAI
+
+client = OpenAI()
+client.api_key = os.environ["OPENAI_API_KEY"]
+
+def extract_git_data(repo_path, author, date):
+    repo = git.Repo(repo_path)
+
+    first_commit = list(repo.iter_commits())[-1].hexsha
+
+    diffs = []
+    author = author if author is not None else get_local_git_author(repo_path)
+    if not author:
+        logging.error(
+            "Author name could not be inferred from local git configuration and was not provided as an argument."
+        )
+        exit(1)
+
+    for commit in repo.iter_commits():
+        commit_date = commit.authored_datetime.date().strftime("%Y-%m-%d")
+        if commit_date == date and commit.author.name == author:
+            diff_data = {
+                "commit_hash": commit.hexsha,
+                "author": commit.author.name,
+                "date": str(commit.authored_datetime),
+                "message": commit.message.strip(),
+                "diff": repo.git.diff(commit.hexsha + "^", commit.hexsha)
+                if first_commit != commit.hexsha
+                else "First Commit",
+            }
+            diffs.append(diff_data)
+
+    return diffs
+
+
+def get_local_git_author(repo_path):
+    try:
+        repo = git.Repo(repo_path, search_parent_directories=True)
+        config_reader = repo.config_reader()
+        author_name = config_reader.get_value("user", "name", None)
+        return author_name
+    except (
+        git.exc.InvalidGitRepositoryError,
+        git.exc.NoSuchPathError,
+        git.exc.GitCommandError,
+        KeyError,
+    ):
+        logging.error(
+            "Could not retrieve the author name from the local git configuration."
+        )
+        return None
+
+
+
+def summarize_diff(diff, model_name="gpt-4"):
+    response = client.chat.completions.create(
+        model=model_name,
+        messages=[
+            {
+                "role": "system",
+                "content": 'Summarize the key changes in this code diff. Don\'t say " the key changes are this" just say the key changes.',
+            },
+            {
+                "role": "user",
+                "content": diff,
+            },
+        ],
+        temperature=1,
+        max_tokens=256,
+        top_p=1,
+        frequency_penalty=0,
+        presence_penalty=0,
+    )
+    return response.choices[0].message.content
+
+
+def summarize_all_diffs(diffs, model_name):
+    # Summarize the diffs
+    summaries = ""
+    for i, diff_data in tqdm(enumerate(diffs, start=1)):
+        diff_summary = summarize_diff(diff_data["diff"], model_name)
+        summaries += f"{i}. Commit: {diff_data['commit_hash']}\n"
+        summaries += f"   Date: {diff_data['date']}\n"
+        summaries += f"   Message: {diff_data['message']}\n"
+        summaries += f"   Key Changes: {diff_summary}\n"
+        summaries += "-" * 40 + "\n"
+    return summaries
+
+
+def daily_summary(summaries, author, date, model_name="gpt-4"):
+    response = client.chat.completions.create(
+        model=model_name,
+        messages=[
+            {
+                "role": "system",
+                "content": f"Create a daily development report for ${author}'s commits on {date}. The recipient of this report is Jack Driscoll. The report should start \"Your day in review\" followed by the top development updates of the day. Don't list by commit, but just give a holistic report of the day.  Here are the key changes made in each commit: ",
+            },
+            {
+                "role": "user",
+                "content": summaries,
+            },
+        ],
+        temperature=1,
+        max_tokens=1000,
+        top_p=1,
+        frequency_penalty=0,
+        presence_penalty=0,
+    )
+    return response.choices[0].message.content
+
+
+def generate(summaries, author, date, model_name):
+    # Summarize the diffs
+    report = daily_summary(summaries, author, date, model_name)
+    print(report)
+    return report
 
 
 def main():
@@ -17,7 +133,7 @@ def main():
     parser.add_argument(
         "--repo",
         help="Path to the repository to generate the report for",
-        required=False,
+        default=".",
     )
     parser.add_argument(
         "--author",
@@ -37,12 +153,14 @@ def main():
     args = parser.parse_args()
     print("Generating daily development report...")
     print("Extracting git data...")
-    diffs = extract_git_data(args.repo, args.author, args.date) 
+    diffs = extract_git_data(args.repo, args.author, args.date)
     print("Summarizing diffs...")
-    summaries = summarize_all_diffs(diffs)
+    summaries = summarize_all_diffs(diffs, args.model)
     print("Generating daily summary...")
-    generate(summaries)
+    report = generate(summaries, args.author, args.date, args.model)
     print("Done!")
+    with open("daily_summary.md", "w") as f:
+        f.write(report)
 
 
 if __name__ == "__main__":
