+++ setup.py
@@ -2,7 +2,7 @@ from setuptools import setup, find_packages
 
 setup(
     name="daily-summary",
-    version="0.1.2",
+    version="0.1.3",
     packages=find_packages(),
     install_requires=[
         # List your project dependencies here
